% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean_space.R
\docType{class}
\name{EuclideanSpace-class}
\alias{EuclideanSpace-class}
\alias{EuclideanSpace}
\title{Euclidean space.}
\description{
Euclidean space.
}
\section{Methods}{

\describe{
\item{\code{Belongs(point)}}{Evaluate if a point belongs to the Euclidean space.}

\item{\code{RandomUniform(n_samples = 1)}}{Sample in the Euclidean space with the uniform distribution.}
}}

