% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invariant_metric.R
\docType{class}
\name{InvariantMetric-class}
\alias{InvariantMetric-class}
\alias{InvariantMetric}
\title{Class for:
   - left-invariant metrics
   - right-invariant metrics
   that exist on Lie groups.
  Points are parameterized by the Riemannian logarithm
   for the canonical left-invariant metric.}
\description{
Class for:
   - left-invariant metrics
   - right-invariant metrics
   that exist on Lie groups.
  Points are parameterized by the Riemannian logarithm
   for the canonical left-invariant metric.
}

