% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manifold.R
\docType{class}
\name{Manifold-class}
\alias{Manifold-class}
\alias{Manifold}
\title{Manifold, i.e. a topological space that locally resembles
Euclidean space near each point.}
\description{
Manifold, i.e. a topological space that locally resembles
Euclidean space near each point.
}

