% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeomstats-package.R
\docType{package}
\name{rgeomstats-package}
\alias{rgeomstats}
\alias{rgeomstats-package}
\title{rgeomstats: An R Interface to Geomstats}
\description{
Provides an R interface to the Python package Geomstats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LMJL-Alea/rgeomstats}
  \item Report bugs at \url{https://github.com/LMJL-Alea/rgeomstats/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@math.cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

}
\keyword{internal}
