% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riemannian_metric.R
\docType{class}
\name{RiemannianMetric-class}
\alias{RiemannianMetric-class}
\alias{RiemannianMetric}
\title{Riemannian and pseudo-Riemannian metrics.}
\description{
Riemannian and pseudo-Riemannian metrics.
}

