% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spd-matrices.R
\name{SPDMatrices}
\alias{SPDMatrices}
\title{Manifold of Symmetric Positive Definite Matrices}
\description{
Class for the manifold of symmetric positive definite (SPD)
matrices.
}
\examples{

## ------------------------------------------------
## Method `SPDMatrices$new`
## ------------------------------------------------

if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  spdm
}

## ------------------------------------------------
## Method `SPDMatrices$belongs`
## ------------------------------------------------

if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  A <- diag(1, 3)
  spdm$belongs(A)
  B <- diag(-1, 3)
  spdm$belongs(B)
}

## ------------------------------------------------
## Method `SPDMatrices$cholesky_factor`
## ------------------------------------------------

if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  V <- cbind(
    c(sqrt(2) / 2, -sqrt(2) / 2, 0),
    c(sqrt(2) / 2, sqrt(2) / 2, 0),
    c(0, 0, 1)
  )
  A <- V \%*\% diag(1:3) \%*\% t(V)
  spdm$cholesky_factor(A)

}
}
\author{
Yann Thanwerdas
}
\concept{discretized curves space}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SPDMatrices-new}{\code{SPDMatrices$new()}}
\item \href{#method-SPDMatrices-belongs}{\code{SPDMatrices$belongs()}}
\item \href{#method-SPDMatrices-cholesky_factor}{\code{SPDMatrices$cholesky_factor()}}
\item \href{#method-SPDMatrices-clone}{\code{SPDMatrices$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SPDMatrices-new"></a>}}
\if{latex}{\out{\hypertarget{method-SPDMatrices-new}{}}}
\subsection{Method \code{new()}}{
The \code{\link{SPDMatrices}} class constructor.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPDMatrices$new(n)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{An integer value representing the shape of the \verb{n x n} matrices.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{\link{SPDMatrices}} \link[R6:R6Class]{R6::R6Class} object.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  spdm
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SPDMatrices-belongs"></a>}}
\if{latex}{\out{\hypertarget{method-SPDMatrices-belongs}{}}}
\subsection{Method \code{belongs()}}{
Check if a matrix is symmetric with positive eigenvalues.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPDMatrices$belongs(mat, atol = 1e-12)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mat}}{A numeric matrix to be checked.}

\item{\code{atol}}{A numeric value specifying the absolute tolerance for
checking. Defaults to \code{1e-12}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A boolean that tells whether the input matrix is SPD.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  A <- diag(1, 3)
  spdm$belongs(A)
  B <- diag(-1, 3)
  spdm$belongs(B)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SPDMatrices-cholesky_factor"></a>}}
\if{latex}{\out{\hypertarget{method-SPDMatrices-cholesky_factor}{}}}
\subsection{Method \code{cholesky_factor()}}{
Compute cholesky factor for a symmetric positive definite
matrix.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPDMatrices$cholesky_factor(mat)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mat}}{An SPD matrix.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The Cholesky factor represented as a numeric vector which stores
its lower triangular matrix including the diagonal elements in a
column-major fashion.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{if (reticulate::py_module_available("geomstats")) {
  spdm <- SPDMatrices$new(n = 3)
  V <- cbind(
    c(sqrt(2) / 2, -sqrt(2) / 2, 0),
    c(sqrt(2) / 2, sqrt(2) / 2, 0),
    c(0, 0, 1)
  )
  A <- V \%*\% diag(1:3) \%*\% t(V)
  spdm$cholesky_factor(A)

}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SPDMatrices-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SPDMatrices-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SPDMatrices$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
