% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgeomstats-package.R
\docType{package}
\name{rgeomstats-package}
\alias{rgeomstats}
\alias{rgeomstats-package}
\title{rgeomstats: Interface to 'Geomstats'}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides an interface to the Python package 'Geomstats' authored by Miolane et al. (2020) \href{https://arxiv.org/abs/2004.04667}{arXiv:2004.04667}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LMJL-Alea/rgeomstats}
  \item \url{https://lmjl-alea.github.io/rgeomstats/}
  \item Report bugs at \url{https://github.com/LMJL-Alea/rgeomstats/issues}
}

}
\author{
\strong{Maintainer}: Aymeric Stamm \email{aymeric.stamm@math.cnrs.fr} (\href{https://orcid.org/0000-0002-8725-3654}{ORCID})

Other contributors:
\itemize{
  \item Nicolas Guigui (\href{https://orcid.org/0000-0002-7901-0732}{ORCID}) (Author of the Geomstats Python package) [contributor]
  \item Alice Le Brigant (\href{https://orcid.org/0000-0002-8055-4753}{ORCID}) (Author of the Geomstats Python package) [contributor]
  \item Johan Mathe (Author of the Geomstats Python package) [contributor]
  \item Nina Miolane (\href{https://orcid.org/0000-0002-1200-9024}{ORCID}) (Author of the Geomstats Python package) [contributor]
  \item Xavier Pennec (\href{https://orcid.org/0000-0002-6617-7664}{ORCID}) (Author of the Geomstats Python package) [contributor]
  \item Luis Pereira (Author of the Geomstats Python package) [contributor]
  \item Yann Thanwerdas (\href{https://orcid.org/0000-0002-9351-6318}{ORCID}) (Author of the Geomstats Python package) [contributor]
}

}
\keyword{internal}
